Revert "Revert "Allow more than INT8_MAX values in LogEvent""
This reverts commit 02205ca0d1ba7f17463090f3265f985f10637326.

Fix out of bounds index issue by making sure mValues isn't empty before
assigning attribution chain indices. Also added a test that fails
without the fix.

Reason for revert: fix added in this CL

Bug: 223273748
Test: atest statsd_test
Test: atest UidAtomTests#TestWriteRawTestAtom
Ignore-AOSP-First: Native StatsEvent APIs are not in AOSP

+++ statsd/src/StatsLogProcessor.cpp
@@ -139,9 +139,9 @@ void StatsLogProcessor::onPeriodicAlarmFired(
 }
 
 void StatsLogProcessor::mapIsolatedUidToHostUidIfNecessaryLocked(LogEvent* event) const {
-    if (std::pair<int, int> indexRange; event->hasAttributionChain(&indexRange)) {
+    if (std::pair<size_t, size_t> indexRange; event->hasAttributionChain(&indexRange)) {
         vector<FieldValue>* const fieldValues = event->getMutableValues();
-        for (int i = indexRange.first; i <= indexRange.second; i++) {
+        for (size_t i = indexRange.first; i <= indexRange.second; i++) {
             FieldValue& fieldValue = fieldValues->at(i);
             if (isAttributionUidField(fieldValue)) {
                 const int hostUid = mUidMap->getHostUidOrSelf(fieldValue.mValue.int_value);
